Pre-Yelp API implementation
- Have not yet connected Yelp API to the SQLite Database
- Need to polish UI
- Possibly add Fragments and Webviews
