+++ app/src/main/java/com/example/stumble/MainActivity.java
@@ -17,6 +17,9 @@ import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -24,7 +27,19 @@ import android.widget.Button;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.io.UnsupportedEncodingException;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
 public class MainActivity extends AppCompatActivity implements View.OnClickListener, SensorEventListener {
+    private static final String API_KEY = "ZOP5mWhWuiDXiaBGjOUHWXVsFCUTDawJ5JFPxKh1D4eth0w7lgdm_AsV_HCwgjXlDv0bagbxctRQK63Y6BOT1a4jlg7jLS1rw173U0AIl11xc-MRTgVaLCiBrj-FY3Yx";
 
     private static final int REQUEST_EVENTS_ACTIVITY = 0;
     private Button eventsPageButton;
@@ -39,7 +54,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private boolean detectShake = false;
 
     private RecyclerView recyclerView;
-    private String listingArray[], descriptionArray[];
 
     private MyDatabase db;
     private SeekBar distanceSeekBar;
@@ -50,6 +64,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     private SharedPreferences sharedPrefs;
     public static final int DEFAULT = 100;
+
+    private float searchLatitude, searchLongitude;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -73,11 +89,11 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         sensorButton = (Button)findViewById(R.id.sensorButton);
         sensorButton.setOnClickListener(this);
 
+        //seekbar
         distanceSeekBar = (SeekBar) findViewById(R.id.distanceSeekBar);
         distanceSeekBar.setOnSeekBarChangeListener(distanceSeekBarListener);
         distanceSeekBar.setMax(200);
-//        int seekBarPreference = sharedPrefs.getInt("Distance", DEFAULT);
-//        distanceSeekBar.setProgress(seekBarPreference);
+        //Saved Preferences/SQLiteDatabase
         savePreferencesButton = (Button)findViewById(R.id.savePreferencesButton);
         savePreferencesButton.setOnClickListener(this);
         loadPreferencesButton = (Button)findViewById(R.id.loadPreferenceButton);
@@ -87,6 +103,13 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
         //Create Database
         db = new MyDatabase(this);
+
+        checkConnection();
+
+        new ReadYelpJSONDataTask().execute(
+                "https://api.yelp.com/v3/businesses/search?latitude="
+                + searchLatitude + "&longitude=" + searchLongitude
+        );
     }
     @Override
     protected void onResume() {
@@ -105,48 +128,61 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     @Override
     public void onClick(View v) {
         Intent i;
+        //List of buttons
         switch (v.getId()) {
             case R.id.eventsPageButton:
+                //Move to events activity
                 i = new Intent(this, EventsActivity.class);
                 getResult.launch(i);
                 break;
             case R.id.diningPageButton:
+                //Move to dining activity
                 i = new Intent(this, DiningActivity.class);
                 getResult.launch(i);
                 break;
             case R.id.dessertsPageButton:
+                //Move to desserts/drinks activity
                 i = new Intent(this, DessertsAndDrinksActivity.class);
                 getResult.launch(i);
                 break;
             case R.id.topPicksPageButton:
+                //Move to top picks activity
                 i = new Intent(this, TopPicksActivity.class);
                 getResult.launch(i);
                 break;
             case R.id.savedPageButton:
+                //Move to saved listings activity
                 i = new Intent(this, SavedActivity.class);
                 getResult.launch(i);
                 break;
             case R.id.sensorButton:
+                //On/off switch for sensor button and detectShake booleans
                 sensorButtonOn = !sensorButtonOn;
                 detectShake = !detectShake;
                 if (sensorButtonOn) {
+                    //If on, set the text (The sensor will start detecting when it is on)
                     sensorButton.setText("Shake phone to receive random event");
                 }
                 else {
+                    //Button is off
                     sensorButton.setText("Random Event");
                 }
                 break;
             case R.id.savePreferencesButton:
+                //Save the preferences
                 sharedPrefs = getSharedPreferences("MyData", Context.MODE_PRIVATE);
                 SharedPreferences.Editor editor = sharedPrefs.edit();
                 editor.putInt("Distance", distanceSeekBar.getProgress());
                 editor.commit();
                 break;
             case R.id.loadPreferenceButton:
+                //Load the preferences
                 int seekBarPreference = sharedPrefs.getInt("Distance", DEFAULT);
                 distanceSeekBar.setProgress(seekBarPreference);
                 break;
             case R.id.SQLiteButton:
+                //Head to SQLite Database Activity (This activity is purely for demo
+                //Will remove this activity later because an API will insert the data instead
                 i = new Intent(this, SQLiteActivity.class);
                 getResult.launch(i);
                 break;
@@ -158,6 +194,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         @Override
         public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
             distanceTextView.setText("Distance (" + progress + "km)");
+            //Tried to implement shared preferences within the seekbar
+            //Did not work as intended...
 //            sharedPrefs = getSharedPreferences("MyData", Context.MODE_PRIVATE);
 //            SharedPreferences.Editor editor = sharedPrefs.edit();
 //            editor.putInt("Distance", progress);
@@ -175,7 +213,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
         }
     };
-
+    //Method for explicit intents
     ActivityResultLauncher<Intent> getResult = registerForActivityResult(
             new ActivityResultContracts.StartActivityForResult(),
             new ActivityResultCallback<ActivityResult>() {
@@ -193,6 +231,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         Sensor sensor = sensorEvent.sensor;
         switch(sensor.getType()) {
             case Sensor.TYPE_ACCELEROMETER:
+                //If the phone is shaken downwards while the sensor is detecting
+                //Move to "Random" Activity (Not random yet)
                 if (Math.abs(sensorEvent.values[1]) > 10.4 && detectShake) {
                     Intent i = new Intent(this, SelectedActivity.class);
                     getResult.launch(i);
@@ -203,5 +243,101 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     @Override
     public void onAccuracyChanged(Sensor sensor, int i) {
 
+    }
+    //Everything below is not implemented yet and will be moved to main activity
+    //This is for getting the API to work
+    public void checkConnection(){
+        ConnectivityManager connectMgr =
+                (ConnectivityManager)getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo networkInfo = connectMgr.getActiveNetworkInfo();
+        if(networkInfo != null && networkInfo.isConnected()){
+            //fetch data
+
+            //display network information in a TextView
+            //String networkInformation = networkInfo.toString();
+            Log.d("debug", "connection ok");
+        }
+        else {
+            //display error
+            Log.d("debug", "network error");
+        }
+    }
+
+    private String readJSONData(String myurl) throws IOException {
+        InputStream is = null;
+        // Only display the first 500 characters of the retrieved
+        // web page content.
+        int len = 2500;
+
+        URL url = new URL(myurl);
+        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+
+        try {
+
+            conn.setReadTimeout(30000 /* milliseconds */);
+            conn.setConnectTimeout(45000 /* milliseconds */);
+            conn.setRequestMethod("GET");
+            conn.setDoInput(true);
+//            conn.setDoOutput(true);
+            conn.setRequestProperty("Authorization", "Bearer " + API_KEY);
+            // Starts the query
+            conn.connect();
+            int response = conn.getResponseCode();
+            Log.d("response2", response +"");
+            if (response == HttpURLConnection.HTTP_OK) {
+                is = conn.getInputStream();
+
+                // Convert the InputStream into a string
+                String contentAsString = readIt(is, len);
+                Log.d("result", contentAsString);
+                return contentAsString;
+            }
+            else {
+                Log.d("response", "Response Code: " + response);
+                return null;
+            }
+
+            // Makes sure that the InputStream is closed after the app is
+            // finished using it.
+        } finally {
+            if (is != null) {
+                is.close();
+                conn.disconnect();
+            }
+        }
+    }
+
+    public String readIt(InputStream stream, int len) throws IOException, UnsupportedEncodingException {
+        Reader reader = null;
+        reader = new InputStreamReader(stream, "UTF-8");
+        char[] buffer = new char[len];
+        reader.read(buffer);
+        return new String(buffer);
+    }
+
+    private class ReadYelpJSONDataTask extends AsyncTask<String, Void, String> {
+
+        Exception exception = null;
+
+        protected String doInBackground(String... urls) {
+            try{
+                Log.d("url", urls[0]);
+                return readJSONData(urls[0]);
+            }catch(IOException e){
+                Log.e("exception", e +"");
+                exception = e;
+            }
+            return null;
+        }
+
+        protected void onPostExecute(String result) {
+            try {
+                Log.d("result", result);
+                JSONObject jsonObject = new JSONObject(result);
+
+            } catch (Exception e) {
+                Log.d("ReadYelpJSONDataTask", e.getLocalizedMessage());
+            }
+        }
     }
 }
