--- app/src/main/java/com/example/stumble/YelpService.java
@@ -1,8 +1,0 @@
-package com.example.stumble;
-
-
-
-public interface YelpService {
-
-
-}
